# packages/types/openapi.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"204":{"description":""}},"tags":["auth"]}},"/api/v1/me":{"get":{"operationId":"UsersController_getMe","parameters":[],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["users"]},"patch":{"operationId":"UsersController_updateMe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMeDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["users"]}},"/api/v1/cafes":{"get":{"operationId":"CafesController_search","parameters":[{"name":"lat","required":false,"in":"query","schema":{"minimum":-90,"maximum":90,"type":"number"}},{"name":"lng","required":false,"in":"query","schema":{"minimum":-180,"maximum":180,"type":"number"}},{"name":"radius","required":false,"in":"query","schema":{"minimum":50,"maximum":5000,"default":1000,"type":"integer"}},{"name":"sort","required":false,"in":"query","schema":{"default":"distance","type":"string","enum":["distance","alphabetical"]}},{"name":"q","required":false,"in":"query","schema":{"minLength":1,"maxLength":100,"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"maximum":9007199254740991,"default":1,"type":"integer"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"integer"}}],"responses":{"200":{"description":""}},"tags":["cafes"]}},"/api/v1/cafes/map":{"get":{"operationId":"CafesController_map","parameters":[{"name":"bbox","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":500,"default":200,"type":"integer"}}],"responses":{"200":{"description":""}},"tags":["cafes"]}},"/api/v1/cafes/{id}":{"get":{"operationId":"CafesController_detail","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"lat","required":false,"in":"query","schema":{"minimum":-90,"maximum":90,"type":"number"}},{"name":"lng","required":false,"in":"query","schema":{"minimum":-180,"maximum":180,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["cafes"]}},"/api/v1/cafes/{id}/discounts":{"get":{"operationId":"CafesController_discounts","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["cafes"]}},"/api/v1/discounts/{id}":{"get":{"operationId":"DiscountsController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["discounts"]},"patch":{"operationId":"DiscountsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDiscountDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["discounts"]},"delete":{"operationId":"DiscountsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"security":[{"access-token":[]}],"tags":["discounts"]}},"/api/v1/cafes/{cafeId}/discounts":{"post":{"operationId":"DiscountsController_create","parameters":[{"name":"cafeId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDiscountDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"tags":["discounts"]}},"/api/v1/cafes/{cafeId}/reports":{"post":{"operationId":"ReportsController_create","parameters":[{"name":"cafeId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReportDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"tags":["reports"]}},"/api/v1/me/reports":{"get":{"operationId":"ReportsController_myReports","parameters":[{"name":"status","required":false,"in":"query","schema":{"type":"string","enum":["pending","reviewing","approved","rejected","auto_registered"]}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["reports"]}},"/api/v1/reports/{id}":{"get":{"operationId":"ReportsController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["reports"]}},"/api/v1/reports/{id}/confirm":{"post":{"operationId":"ReportsController_confirm","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["reports"]},"delete":{"operationId":"ReportsController_cancelConfirm","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["reports"]}},"/api/v1/me/payment-methods":{"get":{"operationId":"PaymentMethodsController_list","parameters":[],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["payment-methods"]},"post":{"operationId":"PaymentMethodsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePaymentMethodDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"tags":["payment-methods"]}},"/api/v1/me/payment-methods/{id}":{"patch":{"operationId":"PaymentMethodsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePaymentMethodDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["payment-methods"]},"delete":{"operationId":"PaymentMethodsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"security":[{"access-token":[]}],"tags":["payment-methods"]}},"/api/v1/cafes/{cafeId}/recommendation":{"get":{"operationId":"RecommendationController_getRecommendation","parameters":[{"name":"cafeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["recommendations"]}},"/api/v1/me/favorites":{"get":{"operationId":"FavoritesController_list","parameters":[],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["favorites"]}},"/api/v1/cafes/{cafeId}/favorites":{"post":{"operationId":"FavoritesController_add","parameters":[{"name":"cafeId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddFavoriteDto"}}}},"responses":{"204":{"description":""}},"security":[{"access-token":[]}],"tags":["favorites"]},"patch":{"operationId":"FavoritesController_updateNotify","parameters":[{"name":"cafeId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFavoriteDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["favorites"]},"delete":{"operationId":"FavoritesController_remove","parameters":[{"name":"cafeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"security":[{"access-token":[]}],"tags":["favorites"]}},"/api/v1/me/payment-alerts":{"get":{"operationId":"PaymentAlertsController_list","parameters":[],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["payment-alerts"]},"post":{"operationId":"PaymentAlertsController_add","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddPaymentAlertDto"}}}},"responses":{"204":{"description":""}},"security":[{"access-token":[]}],"tags":["payment-alerts"]}},"/api/v1/me/payment-alerts/{paymentType}":{"delete":{"operationId":"PaymentAlertsController_remove","parameters":[{"name":"paymentType","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"security":[{"access-token":[]}],"tags":["payment-alerts"]}},"/api/v1/admin/reports":{"get":{"operationId":"AdminReportsController_listQueue","parameters":[{"name":"status","required":false,"in":"query","schema":{"type":"string","enum":["pending","reviewing","approved","rejected","auto_registered"]}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"maximum":9007199254740991,"default":1,"type":"integer"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"integer"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["admin"]}},"/api/v1/admin/reports/{id}":{"get":{"operationId":"AdminReportsController_getReport","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["admin"]}},"/api/v1/admin/reports/{id}/approve":{"post":{"operationId":"AdminReportsController_approve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["admin"]}},"/api/v1/admin/reports/{id}/reject":{"post":{"operationId":"AdminReportsController_reject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RejectReportDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["admin"]}},"/api/v1/admin/crawl-candidates":{"get":{"operationId":"CrawlCandidatesController_listQueue","parameters":[{"name":"status","required":false,"in":"query","schema":{"type":"string","enum":["pending","approved","rejected"]}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"maximum":9007199254740991,"default":1,"type":"integer"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"integer"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["admin"]}},"/api/v1/admin/crawl-candidates/{id}":{"get":{"operationId":"CrawlCandidatesController_getCandidate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["admin"]}},"/api/v1/admin/crawl-candidates/{id}/approve":{"post":{"operationId":"CrawlCandidatesController_approve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApproveCandidateDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["admin"]}},"/api/v1/admin/crawl-candidates/{id}/reject":{"post":{"operationId":"CrawlCandidatesController_reject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RejectCandidateDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["admin"]}},"/api/v1/me/notifications":{"get":{"operationId":"NotificationsController_list","parameters":[{"name":"unread","required":false,"in":"query","schema":{"type":"boolean"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"maximum":9007199254740991,"default":1,"type":"integer"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"integer"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["notifications"]}},"/api/v1/me/notifications/unread-count":{"get":{"operationId":"NotificationsController_unreadCount","parameters":[],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["notifications"]}},"/api/v1/me/notifications/read-all":{"post":{"operationId":"NotificationsController_readAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["notifications"]}},"/api/v1/me/notifications/{id}/read":{"patch":{"operationId":"NotificationsController_markRead","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["notifications"]}},"/api/v1/me/devices":{"get":{"operationId":"DevicesController_list","parameters":[],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["devices"]},"post":{"operationId":"DevicesController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDeviceDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["devices"]}},"/api/v1/me/devices/{id}":{"delete":{"operationId":"DevicesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"tags":["devices"]}},"/api/v1/uploads/receipt-presign":{"post":{"operationId":"UploadsController_presignReceipt","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresignReceiptDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"tags":["uploads"]}},"/api/v1/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":""}},"tags":["health"]}}},"info":{"title":"Catch Coffee API","description":"카페 할인/결제수단 추천 서비스 API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignupDto":{"type":"object","properties":{"email":{"type":"string","maxLength":255,"format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"},"password":{"type":"string","minLength":8,"maxLength":72},"name":{"type":"string","minLength":1,"maxLength":100},"phone":{"type":"string","maxLength":20}},"required":["email","password","name"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","maxLength":255,"format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"},"password":{"type":"string","minLength":1,"maxLength":72}},"required":["email","password"]},"RefreshDto":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"UpdateMeDto":{"type":"object","properties":{"name":{"type":"string","minLength":1,"maxLength":100},"phone":{"type":"string","maxLength":20,"nullable":true}}},"CreateDiscountDto":{"type":"object","properties":{"title":{"type":"string","minLength":1,"maxLength":200},"discountType":{"type":"string","enum":["percentage","amount"]},"discountValue":{"type":"number","exclusiveMinimum":true,"minimum":0},"targetScope":{"default":"all","type":"string","enum":["all","menu"]},"conditions":{"type":"object","additionalProperties":{}},"paymentType":{"type":"string","enum":["naverpay","kakaopay","card","other"]},"startAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"},"endAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"}},"required":["title","discountType","discountValue"]},"UpdateDiscountDto":{"type":"object","properties":{"title":{"type":"string","minLength":1,"maxLength":200},"discountType":{"type":"string","enum":["percentage","amount"]},"discountValue":{"type":"number","exclusiveMinimum":true,"minimum":0},"targetScope":{"type":"string","enum":["all","menu"]},"conditions":{"type":"object","additionalProperties":{}},"paymentType":{"type":"string","enum":["naverpay","kakaopay","card","other"]},"startAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"},"endAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"},"status":{"type":"string","enum":["scheduled","active","ended","hidden"]}}},"CreateReportDto":{"type":"object","properties":{"title":{"type":"string","minLength":1,"maxLength":200},"discountType":{"type":"string","enum":["percentage","amount"]},"discountValue":{"type":"number","exclusiveMinimum":true,"minimum":0},"conditions":{"type":"object","additionalProperties":{}},"infoSource":{"type":"string","enum":["offline","receipt","store_notice","witnessed"]},"receiptImageUrl":{"type":"string","maxLength":2048,"format":"uri"},"startAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"},"endAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"}},"required":["title","discountType","discountValue","infoSource","receiptImageUrl"]},"CreatePaymentMethodDto":{"type":"object","properties":{"type":{"type":"string","enum":["naverpay","kakaopay","card","other"]},"label":{"type":"string","minLength":1,"maxLength":100},"token":{"type":"string","minLength":1,"maxLength":2048},"isDefault":{"type":"boolean"}},"required":["type","label"]},"UpdatePaymentMethodDto":{"type":"object","properties":{"label":{"type":"string","minLength":1,"maxLength":100},"token":{"type":"string","minLength":1,"maxLength":2048},"isDefault":{"type":"boolean"}}},"AddFavoriteDto":{"type":"object","properties":{"notifyEnabled":{"type":"boolean"}}},"UpdateFavoriteDto":{"type":"object","properties":{"notifyEnabled":{"type":"boolean"}},"required":["notifyEnabled"]},"AddPaymentAlertDto":{"type":"object","properties":{"paymentType":{"type":"string","enum":["naverpay","kakaopay","card","other"]}},"required":["paymentType"]},"RejectReportDto":{"type":"object","properties":{"reason":{"type":"string","minLength":1,"maxLength":500}},"required":["reason"]},"ApproveCandidateDto":{"type":"object","properties":{"cafeId":{"type":"string","minLength":1},"title":{"type":"string","minLength":1,"maxLength":200},"discountType":{"type":"string","enum":["percentage","amount"]},"discountValue":{"type":"number","exclusiveMinimum":true,"minimum":0}}},"RejectCandidateDto":{"type":"object","properties":{"reason":{"type":"string","maxLength":500}}},"RegisterDeviceDto":{"type":"object","properties":{"expoPushToken":{"type":"string","minLength":1,"maxLength":255},"platform":{"type":"string","enum":["ios","android","web"]}},"required":["expoPushToken","platform"]},"PresignReceiptDto":{"type":"object","properties":{"contentType":{"type":"string","enum":["image/jpeg","image/png","image/webp"]}},"required":["contentType"]}}}}